2020,3,25

+++ app/src/main/java/com/fly/ui/UIView.kt
@@ -49,7 +49,7 @@ class UIView(context : Context?,attrs: AttributeSet?) : View(context,attrs)
         return true
     }
 
-    private fun Collision(x1:Int,y1:Int,width1:Int,height1:Int,x2:Int,y2:Int,width2:Int,height2:Int) : Boolean
+    private fun Collision(x1:Float,y1:Float,width1:Float,height1:Float,x2:Float,y2:Float,width2:Float,height2:Float) : Boolean
     {
         if (x1 <= x2 + width2 && x1 + width1 >= x2 && y1 + height1 >= y2 && y1 <= y2 + height2)
             return true
@@ -65,13 +65,13 @@ class UIView(context : Context?,attrs: AttributeSet?) : View(context,attrs)
     fun AddWidget(widget:Widget) { widget_list.add(widget) }
     fun AddWidgets(widgets: ArrayList<Widget>) { widget_list.addAll(widgets) }
 
-    fun FindWidget(x:Int,y:Int) : ArrayList<Int>
+    fun FindWidget(x:Float,y:Float) : ArrayList<Int>
     {
         val index : ArrayList<Int> = ArrayList<Int>()
         for (i in 0 until widget_list.size)
         {
             if (Collision(widget_list[i].x,widget_list[i].y,widget_list[i].width,widget_list[i].height,
-                    x,y,1,1))
+                    x,y,1f,1f))
             {
                 index.add(i)
             }
@@ -81,7 +81,7 @@ class UIView(context : Context?,attrs: AttributeSet?) : View(context,attrs)
 
     fun WidgetClick(event: MotionEvent)
     {
-        val index = FindWidget(event.x.toInt(),event.y.toInt())
+        val index = FindWidget(event.x,event.y)
         for (i in 0 until index.size)
         {
             widget_list[index[i]].is_up = false
